@@ -3,7 +3,7 @@ import {ConfigService} from 'src/app/_services/config.service';
 import {MapService} from '../../_services/map.service';
 import { CustomIconService } from '../../_services/custom-icon.service';
 import {IBasemap, IPageFormat} from 'src/app/_models/IConfig';
-import {FormControl, FormGroup} from '@angular/forms';
+import {UntypedFormControl, UntypedFormGroup} from '@angular/forms';
 import {debounceTime, switchMap} from 'rxjs/operators';
 import Geometry from 'ol/geom/Geometry';
 import {MatAutocompleteSelectedEvent} from '@angular/material/autocomplete';
@@ -64,8 +64,8 @@ export class MapComponent implements OnInit {
   activeTab = 0;
 
   // Geocoder
-  formGeocoder = new FormGroup({
-    search: new FormControl('')
+  formGeocoder = new UntypedFormGroup({
+    search: new UntypedFormControl('')
   });
   geocoderGroupOptions: {
     id: string;
